+++ src/authenticate/decode.js
