+++ src/authenticate/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "authenticate",
+  "main": "index.js"
+}
