+++ src/gradebook/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "gradebook",
+  "main": "index.js"
+}
